fix: 修复文档状态数据数据错误问题
