+++ apps/common/util/page_utils.py
@@ -18,10 +18,11 @@ def page(query_set, page_size, handler, is_the_task_interrupted=lambda: False):
     @param is_the_task_interrupted: 任务是否被中断
     @return:
     """
+    query = query_set.order_by("id")
     count = query_set.count()
     for i in range(0, ceil(count / page_size)):
         if is_the_task_interrupted():
             return
         offset = i * page_size
-        paragraph_list = query_set[offset: offset + page_size]
+        paragraph_list = query.all()[offset: offset + page_size]
         handler(paragraph_list)
