+++ ui/src/views/document/component/Status.vue
@@ -1,51 +1,13 @@
 <template>
-  <el-popover placement="top" :width="450" trigger="hover">
-    <template #default>
-      <el-row :gutter="3" v-for="status in statusTable" :key="status.type">
-        <el-col :span="4">{{ taskTypeMap[status.type] }} </el-col>
-        <el-col :span="4">
-          <el-text v-if="status.state === State.SUCCESS || status.state === State.REVOKED">
-            <el-icon class="success"><SuccessFilled /></el-icon>
-            {{ stateMap[status.state](status.type) }}
-          </el-text>
-          <el-text v-else-if="status.state === State.FAILURE">
-            <el-icon class="danger"><CircleCloseFilled /></el-icon>
-            {{ stateMap[status.state](status.type) }}
-          </el-text>
-          <el-text v-else-if="status.state === State.STARTED">
-            <el-icon class="is-loading primary"><Loading /></el-icon>
-            {{ stateMap[status.state](status.type) }}
-          </el-text>
-          <el-text v-else-if="status.state === State.PENDING">
-            <el-icon class="is-loading primary"><Loading /></el-icon>
-            {{ stateMap[status.state](status.type) }}
-          </el-text>
-          <el-text v-else-if="aggStatus?.value === State.REVOKE">
-            <el-icon class="is-loading primary"><Loading /></el-icon>
-            {{ stateMap[aggStatus.value](aggStatus.key) }}
-          </el-text>
-        </el-col>
-        <el-col :span="5">
-          完成
-          {{
-            Object.keys(status.aggs ? status.aggs : {})
-              .filter((k) => k == State.SUCCESS)
-              .map((k) => status.aggs[k])
-              .reduce((x: any, y: any) => x + y, 0)
-          }}/{{
-            Object.values(status.aggs ? status.aggs : {}).reduce((x: any, y: any) => x + y, 0)
-          }}
-        </el-col>
-        <el-col :span="9">
-          {{
-            status.time
-              ? status.time[
-                  status.state == State.REVOKED ? State.REVOKED : State.PENDING
-                ]?.substring(0, 19)
-              : undefined
-          }}
-        </el-col>
-      </el-row>
+  <el-popover v-model:visible="visible" placement="top" :width="450" trigger="hover">
+    <template #default
+      ><StatusTable
+        v-if="visible"
+        :status="status"
+        :statusMeta="statusMeta"
+        :taskTypeMap="taskTypeMap"
+        :stateMap="stateMap"
+      ></StatusTable>
     </template>
     <template #reference>
       <el-text v-if="aggStatus?.value === State.SUCCESS || aggStatus?.value === State.REVOKED">
@@ -72,11 +34,11 @@
   </el-popover>
 </template>
 <script setup lang="ts">
-import { computed } from 'vue'
-import { Status, TaskType, State, type TaskTypeInterface } from '@/utils/status'
-import { mergeWith } from 'lodash'
+import { computed, ref } from 'vue'
+import { TaskType, State } from '@/utils/status'
+import StatusTable from '@/views/document/component/StatusTable.vue'
 const props = defineProps<{ status: string; statusMeta: any }>()
-
+const visible = ref<boolean>(false)
 const checkList: Array<string> = [
   State.REVOKE,
   State.STARTED,
@@ -112,56 +74,5 @@ const stateMap: any = {
   [State.FAILURE]: (type: number) => '失败',
   [State.SUCCESS]: (type: number) => '成功'
 }
-
-const parseAgg = (agg: { count: number; status: string }) => {
-  const status = new Status(agg.status)
-  return Object.keys(TaskType)
-    .map((key) => {
-      const value = TaskType[key as keyof TaskTypeInterface]
-      return { [value]: { [status.task_status[value]]: agg.count } }
-    })
-    .reduce((x, y) => ({ ...x, ...y }), {})
-}
-
-const customizer: (x: any, y: any) => any = (objValue: any, srcValue: any) => {
-  if (objValue == undefined && srcValue) {
-    return srcValue
-  }
-  if (srcValue == undefined && objValue) {
-    return objValue
-  }
-  // 如果是数组，我们将元素进行聚合
-  if (typeof objValue === 'object' && typeof srcValue === 'object') {
-    // 若是object类型的对象，我们进行递归
-    return mergeWith(objValue, srcValue, customizer)
-  } else {
-    // 否则，单纯的将值进行累加
-    return objValue + srcValue
-  }
-}
-const aggs = computed(() => {
-  return (props.statusMeta.aggs ? props.statusMeta.aggs : [])
-    .map((agg: any) => {
-      return parseAgg(agg)
-    })
-    .reduce((x: any, y: any) => {
-      return mergeWith(x, y, customizer)
-    }, {})
-})
-
-const statusTable = computed(() => {
-  return Object.keys(TaskType)
-    .map((key) => {
-      const value = TaskType[key as keyof TaskTypeInterface]
-      const parseStatus = new Status(props.status)
-      return {
-        type: value,
-        state: parseStatus.task_status[value],
-        aggs: aggs.value[value],
-        time: props.statusMeta.state_time[value]
-      }
-    })
-    .filter((item) => item.state !== State.IGNORED)
-})
 </script>
 <style lang="scss" scoped></style>
